renamed listeners: private folder to innerCirclew
